+++ scalus-examples/jvm/src/test/scala/scalus/examples/betting/BettingTransactionTest.scala
@@ -289,7 +289,7 @@ class BettingTransactionTest extends AnyFunSuite, ScalusTest {
         val result = runValidator(provider, joinTx, betUtxo.input)
         assert(result.isSuccess)
         assert(
-          result.budget == (ExUnits(memory = 115515, steps = 45_440080))
+          result.budget == (ExUnits(memory = 115215, steps = 45_392080))
         )
 
         provider.setSlot(beforeSlot - 1)
@@ -514,7 +514,7 @@ class BettingTransactionTest extends AnyFunSuite, ScalusTest {
         val result = runValidator(provider, winTx, joinedBetUtxo.input)
         assert(result.isSuccess)
         assert(
-          result.budget == (ExUnits(memory = 103398, steps = 39_995168))
+          result.budget == (ExUnits(memory = 103098, steps = 39_947168))
         )
 
         provider.setSlot(env.slotConfig.timeToSlot(afterTime))