+++ scalus-examples/jvm/src/test/scala/scalus/examples/editablenft/EditableNftValidatorTest.scala
@@ -558,7 +558,7 @@ class EditableNftValidatorTest extends AnyFunSuite, ScalusTest {
           signer = Alice.signer
         )
         assertResult(
-          ExUnits(memory = 73790, steps = 25_179592)
+          ExUnits(memory = 73190, steps = 25_083592)
         ):
             burnTx.witnessSet.redeemers.get.value.totalExUnits
         val burnResult = provider.submit(burnTx).await()