+++ scalus-examples/jvm/src/test/scala/scalus/examples/escrow/EscrowTest.scala
@@ -109,7 +109,7 @@ class EscrowTest extends AnyFunSuite, ScalusTest {
         )
 
         assertResult(
-          ExUnits(memory = 103013, steps = 50_831003)
+          ExUnits(memory = 102713, steps = 50_783003)
         ):
             depositTx.witnessSet.redeemers.get.value.totalExUnits
         val result = provider.submit(depositTx).await()
@@ -214,7 +214,7 @@ class EscrowTest extends AnyFunSuite, ScalusTest {
         )
 
         assertResult(
-          ExUnits(memory = 95636, steps = 41_985697)
+          ExUnits(memory = 95336, steps = 41_937697)
         ):
             payTx.witnessSet.redeemers.get.value.totalExUnits
         val result = provider.submit(payTx).await()
@@ -299,7 +299,7 @@ class EscrowTest extends AnyFunSuite, ScalusTest {
         )
 
         assertResult(
-          ExUnits(memory = 106609, steps = 48_872278)
+          ExUnits(memory = 102784, steps = 46_412776)
         ):
             refundTx.witnessSet.redeemers.get.value.totalExUnits
         val result = provider.submit(refundTx).await()