+++ scalus-examples/jvm/src/test/scala/scalus/examples/htlc/HtlcTest.scala
@@ -61,7 +61,7 @@ class HtlcTest extends AnyFunSuite, ScalusTest {
     }
 
     test(s"HTLC validator size is ${HtlcContract.compiled.script.script.size} bytes") {
-        assert(HtlcContract.compiled.script.script.size == 318)
+        assert(HtlcContract.compiled.script.script.size == 315)
     }
 
     test("VALIDATOR: receiver reveals preimage before timeout") {
@@ -84,8 +84,8 @@ class HtlcTest extends AnyFunSuite, ScalusTest {
         assert(Try(contract(scriptCtx.toData).code).isSuccess)
         val result = contract(scriptCtx.toData).program.evaluateDebug
         assert(result.isSuccess)
-        assert(result.budget == ExUnits(memory = 27127, steps = 11_541882))
-        assert(result.budget.fee == Coin(2398))
+        assert(result.budget == ExUnits(memory = 26827, steps = 11_493882))
+        assert(result.budget.fee == Coin(2377))
     }
 
     test("receiver reveals preimage before timeout") {
@@ -104,7 +104,7 @@ class HtlcTest extends AnyFunSuite, ScalusTest {
           signer = Bob.signer
         )
 
-        assertResult(ExUnits(memory = 27127, steps = 11_541882)):
+        assertResult(ExUnits(memory = 26827, steps = 11_493882)):
             revealTx.witnessSet.redeemers.get.value.totalExUnits
 
         provider.setSlot(beforeSlot)
@@ -185,7 +185,7 @@ class HtlcTest extends AnyFunSuite, ScalusTest {
           signer = Alice.signer
         )
 
-        assertResult(ExUnits(memory = 24194L, steps = 8732711L)):
+        assertResult(ExUnits(memory = 23894L, steps = 8684711L)):
             timeoutTx.witnessSet.redeemers.get.value.totalExUnits
 
         provider.setSlot(afterSlot)