+++ scalus-examples/jvm/src/test/scala/scalus/examples/lottery/LotteryValidatorTest.scala
@@ -46,7 +46,7 @@ class LotteryValidatorTest extends AnyFunSuite, ScalusTest {
           provider,
           revealTx,
           lotteryUtxo._1,
-          ExUnits(memory = 67953, steps = 26_969150)
+          ExUnits(memory = 67653, steps = 26_921150)
         )
     }
 
@@ -73,7 +73,7 @@ class LotteryValidatorTest extends AnyFunSuite, ScalusTest {
           provider,
           revealTx,
           lotteryUtxo._1,
-          ExUnits(memory = 68185, steps = 27_037637)
+          ExUnits(memory = 67885, steps = 26_989637)
         )
     }
 
@@ -100,7 +100,7 @@ class LotteryValidatorTest extends AnyFunSuite, ScalusTest {
           provider,
           revealTx,
           lotteryUtxo._1,
-          ExUnits(memory = 67953, steps = 26_969150)
+          ExUnits(memory = 67653, steps = 26_921150)
         )
     }
 
@@ -241,7 +241,7 @@ class LotteryValidatorTest extends AnyFunSuite, ScalusTest {
           provider,
           p2RevealTx,
           lotteryUtxo2._1,
-          ExUnits(memory = 47631, steps = 18_066134)
+          ExUnits(memory = 47331, steps = 18_018134)
         )
     }
 
@@ -288,7 +288,7 @@ class LotteryValidatorTest extends AnyFunSuite, ScalusTest {
           provider,
           timeoutTx,
           lotteryUtxo2._1,
-          ExUnits(memory = 51535, steps = 19_380131)
+          ExUnits(memory = 51235, steps = 19_332131)
         )
     }
 
@@ -376,7 +376,7 @@ class LotteryValidatorTest extends AnyFunSuite, ScalusTest {
           provider,
           loseTx,
           lotteryUtxo2._1,
-          ExUnits(memory = 49481, steps = 19_741679)
+          ExUnits(memory = 49181, steps = 19_693679)
         )
     }
 
@@ -421,7 +421,7 @@ class LotteryValidatorTest extends AnyFunSuite, ScalusTest {
           provider,
           timeoutTx,
           lotteryUtxo2._1,
-          ExUnits(memory = 51303, steps = 19_311644)
+          ExUnits(memory = 51003, steps = 19_263644)
         )
     }
 
@@ -571,7 +571,7 @@ class LotteryValidatorTest extends AnyFunSuite, ScalusTest {
           provider,
           loseTx,
           lotteryUtxo2._1,
-          ExUnits(memory = 49249, steps = 19_673192)
+          ExUnits(memory = 48949, steps = 19_625192)
         )
     }
 
@@ -741,7 +741,7 @@ class LotteryValidatorTest extends AnyFunSuite, ScalusTest {
           provider,
           p1RevealTx,
           lotteryUtxo2._1,
-          ExUnits(memory = 47399, steps = 17_997647)
+          ExUnits(memory = 47099, steps = 17_949647)
         )
     }
 }