+++ scalus-examples/jvm/src/test/scala/scalus/examples/vesting/VestingTransactionTest.scala
@@ -139,7 +139,7 @@ class VestingTransactionTest extends AnyFunSuite, ScalusTest {
         val result = runValidator(provider, withdrawTx, lockedUtxo.input)
         assert(result.isSuccess, s"Validator failed: $result")
         assert(
-          result.budget == (ExUnits(memory = 103552, steps = 45_690020))
+          result.budget == (ExUnits(memory = 103252, steps = 45_642020))
         )
 
         val submitResult = provider.submit(withdrawTx).await()
@@ -172,7 +172,7 @@ class VestingTransactionTest extends AnyFunSuite, ScalusTest {
         val result = runValidator(provider, withdrawTx, lockedUtxo.input)
         assert(result.isSuccess, s"Validator failed: $result")
         assert(
-          result.budget == (ExUnits(memory = 127895, steps = 62_482008))
+          result.budget == (ExUnits(memory = 127595, steps = 62_434008))
         )
 
         val submitResult = provider.submit(withdrawTx).await()